fixup! Router worker: send SIGINT when the parent process terminates
