+++ packages/next/src/server/utils.ts
@@ -72,12 +72,3 @@ export function debounce<T, F extends AnyFunc<T>>(
     }
   }
 }
-
-export function hasRenderWorker() {
-  return Boolean(
-    typeof jest === 'undefined' &&
-      !process.env.NEXT_PRIVATE_WORKER &&
-      (process.env.__NEXT_PRIVATE_PREBUNDLED_REACT ||
-        process.env.NODE_ENV === 'development')
-  )
-}
